Pampo

+++ app/Http/Controllers/CursoController.php
@@ -7,10 +7,10 @@ use App\Models\Curso;
 use Illuminate\Support\Facades\Auth;
 
 
+
 class CursoController extends Controller
 {
     public function index(){
-
 
         $search = request('search');
 
@@ -24,6 +24,7 @@ class CursoController extends Controller
         }
         return view('cursos.index',['cursos' => $cursos,'search' => $search]);
     }
+
 
     public function show($id){
         $curso = curso::findOrfail($id);
@@ -45,7 +46,9 @@ class CursoController extends Controller
 
     public function join($id){
         $user = Auth::user();
-
+        if($user->perm != 0){
+            return reditect ('/cursos');
+        }
         $cursosusuario = $user->cursos->toArray();
         foreach ($cursosusuario as $cursousuario){
             if($cursousuario['id'] == $id){
@@ -56,15 +59,18 @@ class CursoController extends Controller
 
         $curso = curso::findOrfail($id);
 
-        return redirect ('/home')->with('msg','Matriculado com sucesso em' . $curso->name);
+        return redirect ('/home')->with('status','Matriculado com sucesso em: ' . $curso->name);
     }
 
     public function leave($id){
+        if($user->perm != 0){
+            return reditect ('/cursos');
+        }
         $user = Auth::user();
 
         $user->cursos()->detach($id);
         $curso = curso::findOrfail($id);
 
-        return redirect ('/home')->with('msg','Você se desmatriculou de:' . $curso->name);
+        return redirect ('/home')->with('status','Você se desmatriculou de: ' . $curso->name);
     }
 }
